quantumHardwareJobs/python_requirements.txt
amazon-braket-default-simulator==1.23.1
amazon-braket-schemas==1.21.4
amazon-braket-sdk==1.78.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.2
appnope==0.1.4
asttokens==2.4.1
backoff==2.2.1
backports.entry-points-selectable==1.3.0
boltons==24.0.0
boto3==1.34.95
botocore==1.34.95
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fonttools==4.51.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openpulse==0.5.0
openqasm3==0.5.0
opt-einsum==3.3.0
oqpy==0.3.5
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
s3transfer==0.10.1
scipy==1.13.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
